--- constants.ts
@@ -1,19 +1,0 @@
-export const musicFiles = {
-  'victory': new URL('./audio/11L-victory_music-1756395358872.mp3', import.meta.url).href,
-  'country': new URL('./audio/It\'s a Spelling Bee! (Country).mp3', import.meta.url).href,
-  'country_instrumental': new URL('./audio/It\'s a Spelling Bee! (Country Instrumental).mp3', import.meta.url).href,
-  'deep_bass_instrumental': new URL('./audio/It\'s a Spelling Bee! (Deep Bass Instrumental).mp3', import.meta.url).href,
-  'deep_bass': new URL('./audio/It\'s a Spelling Bee! (Deep Bass).mp3', import.meta.url).href,
-  'funk_instrumental': new URL('./audio/It\'s a Spelling Bee! (Funk Instrumental).mp3', import.meta.url).href,
-  'funk': new URL('./audio/It\'s a Spelling Bee! (Funk).mp3', import.meta.url).href,
-  'latin_instrumental': new URL('./audio/It\'s a Spelling Bee! (Latin Instrumental).mp3', import.meta.url).href,
-  'latin': new URL('./audio/It\'s a Spelling Bee! (Latin).mp3', import.meta.url).href,
-  'rock_instrumental': new URL('./audio/It\'s a Spelling Bee! (Rock Instrumental).mp3', import.meta.url).href,
-  'rock': new URL('./audio/It\'s a Spelling Bee! (Rock).mp3', import.meta.url).href,
-  'spooky_instrumental': new URL('./audio/It\'s a Spelling Bee! (Spooky Instrumental).mp3', import.meta.url).href,
-  'spooky': new URL('./audio/It\'s a Spelling Bee! (Spooky).mp3', import.meta.url).href,
-  'tech_house_instrumental': new URL('./audio/It\'s a Spelling Bee! (Tech House Instrumental).mp3', import.meta.url).href,
-  'tech_house': new URL('./audio/It\'s a Spelling Bee! (Tech House).mp3', import.meta.url).href,
-};
-
-export const DEFAULT_STYLE = 'funk';
