--- index.html.bak
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="theme-color" content="#4f46e5">
-    <meta name="mobile-web-app-capable" content="yes">
-    <meta name="apple-mobile-web-app-status-bar-style" content="default">
-    <title>Interactive Spelling Bee Championship</title>
-
-    <link rel="stylesheet" href="/spelling-bee-game/tailwind.css">
-
-    <link rel="manifest" href="/spelling-bee-game/manifest.webmanifest">
-    <link rel="icon" href="/spelling-bee-game/icons/icon.svg" type="image/svg+xml">
-    <link rel="apple-touch-icon" href="/spelling-bee-game/icons/icon.svg">
-    <link rel="stylesheet" href="/spelling-bee-game/style.css">
-</head>
-<body>
-    <div id="root"></div>
-    
-    <script type="module" src="/spelling-bee-game/app.js"></script>
-    <!--
-    <script>
-        if ('serviceWorker' in navigator) {
-            window.addEventListener('load', () => {
-                navigator.serviceWorker.register('service-worker.js');
-            });
-        }
-    </script>
-    -->
-    <!-- Service worker temporarily disabled for debugging -->
-    <script>
-        document.addEventListener('click', (e) => {
-            const target = e.target;
-            if (target instanceof HTMLButtonElement) {
-                if (window.matchMedia('(prefers-reduced-motion: reduce)').matches) return;
-                target.classList.add('btn-bounce');
-                target.addEventListener('animationend', () => {
-                    target.classList.remove('btn-bounce');
-                }, { once: true });
-            }
-        });
-    </script>
-</body>
-</html><!-- Force rebuild -->
-<!-- Force rebuild -->
-<!-- Force update -->
-<!-- Final update -->
-<!-- Latest fix -->
-<!-- Final deployment fix -->
