+++ src/ResultsScreen.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { GameResults, GameConfig, LeaderboardEntry } from './types';
-import applauseSoundFile from './audio/applause.mp3';
+import applauseSoundFile from '../audio/applause.mp3';
 import { launchConfetti } from './utils/confetti';
 import { recordDailyCompletion, StreakInfo } from './DailyChallenge';
-import beeImg from './img/avatars/bee.svg';
+import beeImg from '../img/avatars/bee.svg';
 
 interface ResultsScreenProps {
   results: GameResults;
