+++ src/confetti.ts
