+++ src/utils/confetti.ts
